+++ tencentcloud/resource_tc_kubernetes_scale_worker_test.go
@@ -141,12 +141,9 @@ variable "availability_zone" {
   default = "ap-guangzhou-3"
 }
 
-variable "vpc" {
-  default = "` + defaultVpcId + `"
-}
-
-variable "subnet" {
-  default = "` + defaultSubnetId + `"
+data "tencentcloud_vpc_subnets" "vpc" {
+    is_default        = true
+    availability_zone = var.availability_zone
 }
 
 variable "default_instance_type" {
@@ -157,7 +154,7 @@ variable "scale_instance_type" {
   default = "S2.LARGE16"
 }
 resource "tencentcloud_kubernetes_cluster" "managed_cluster" {
-  vpc_id                  = var.vpc
+  vpc_id                  = data.tencentcloud_vpc_subnets.vpc.instance_list.0.vpc_id
   cluster_cidr            = "192.168.0.0/16"
   cluster_max_pod_num     = 32
   cluster_name            = "test"
@@ -173,7 +170,7 @@ resource "tencentcloud_kubernetes_cluster" "managed_cluster" {
     internet_charge_type       = "TRAFFIC_POSTPAID_BY_HOUR"
     internet_max_bandwidth_out = 100
     public_ip_assigned         = true
-    subnet_id                  = var.subnet
+    subnet_id                  = data.tencentcloud_vpc_subnets.vpc.instance_list.0.subnet_id
 
     data_disk {
       disk_type = "CLOUD_PREMIUM"
@@ -199,7 +196,7 @@ resource tencentcloud_kubernetes_scale_worker test_scale {
 	instance_charge_type	   				= "PREPAID"
 	instance_charge_type_prepaid_period 	= 6
 	instance_charge_type_prepaid_renew_flag = "NOTIFY_AND_AUTO_RENEW"
-    subnet_id                  				= var.subnet
+    subnet_id                  				= data.tencentcloud_vpc_subnets.vpc.instance_list.0.subnet_id
     system_disk_type           				= "CLOUD_SSD"
     system_disk_size           				= 50
     internet_charge_type       				= "TRAFFIC_POSTPAID_BY_HOUR"